-Implemented H264 Video Writing in RTSP Server

+++ NetworkManagerAPI/StringPackage.cc
@@ -1,6 +1,6 @@
 #include "StringPackage.h"
 
-StringPackage::StringPackage(uint16_t max_size) :
+StringPackage::StringPackage(size_t max_size) :
 	mSmartString{ std::make_unique<char[]>(max_size) },
 	mMaxSize{ max_size }
 {
@@ -12,17 +12,17 @@ const char* StringPackage::cData() const
 	return mSmartString.get();
 }
 
-uint16_t StringPackage::getCurrentSize() const
+size_t StringPackage::getCurrentSize() const
 {
 	return mCurrentSize;
 }
 
-uint16_t StringPackage::getMaxSize() const
+size_t StringPackage::getMaxSize() const
 {
 	return mMaxSize;
 }
 
-void StringPackage::setCurrentSize(uint16_t curr_size)
+void StringPackage::setCurrentSize(size_t curr_size)
 {
 	mCurrentSize = curr_size;
 }
