+++ RtspClient/include/rtp_frame_package.h
@@ -12,18 +12,18 @@
 class RtpFramePackage : public IPackage
 {
 public:
-	RtpFramePackage(uint16_t size = RTP_HEADER_SIZE + MAX_RTP_PAYLOAD_SIZE);
+	RtpFramePackage(size_t size = RTP_HEADER_SIZE + MAX_RTP_PAYLOAD_SIZE);
 	~RtpFramePackage() = default;
 
 	const char* cData() const override;
 	char* data() override;
 
-	uint16_t getCurrentSize() const override;
-	uint16_t getMaxSize() const override;
+	size_t getCurrentSize() const override;
+	size_t getMaxSize() const override;
 
 	const RtpHeader* getRtpHeader() const;
 private:
-	uint16_t mSize;
+	size_t mSize;
 	std::unique_ptr<uint8_t[]> mData;
 	RtpHeader* mHeaderPtr;
 };
